Refactor Renderer into HtmlRenderer and JsonRenderer
Created a base meta class that defines the interface of a Renderer that
all Renderers inherit off of. Site now has a `render()` method which is
what calls all renderers, rather than having the CLI do it.

Also added an empty RssRenderer and AtomRenderer which will be filled in
later. Also added config to support these.
