+++ posty/config.py
@@ -60,5 +60,9 @@ class Config(MutableMapping):
         c.setdefault('num_top_tags', 5)
         c.setdefault('num_posts_per_page', 5)
 
+        c.setdefault('feeds', {})
+        c['feeds'].setdefault('rss', True)
+        c['feeds'].setdefault('atom', True)
+
         c.setdefault('compat', {})
         c['compat'].setdefault('redirect_posty1_urls', False)
