+++ posty/renderer/atom.py
@@ -1,0 +1,5 @@
+from .base import Renderer
+
+
+class AtomRenderer(Renderer):
+    pass
