+++ posty/renderer/html.py
@@ -3,11 +3,12 @@ install_aliases()   # noqa
 
 from collections import defaultdict
 import jinja2
-import json
 import os
 from urllib.parse import urljoin
 
-from . import template_filters, util
+from .. import util
+from .base import Renderer
+from .util import markdown, media_url_func
 
 # Route reference
 # /               Posts
@@ -20,7 +21,7 @@ from . import template_filters, util
 # /:slug/         Page matching :slug
 
 
-class Renderer(object):
+class HtmlRenderer(Renderer):
     def __init__(self, site, output_path='build'):
         """
         :param site:
@@ -29,26 +30,16 @@ class Renderer(object):
         :param output_path:
             path relative to the Site's path to put rendered HTML files into
         """
-        self.site = site
-        self.output_path = os.path.join(site.site_path, output_path)
+        super(HtmlRenderer, self).__init__(site, output_path=output_path)
 
         template_path = os.path.join(site.site_path, 'templates')
         self.jinja_env = jinja2.Environment(
             loader=jinja2.FileSystemLoader(template_path),
         )
-        self.set_jinja_filters()
 
-    def set_jinja_filters(self):
-        """
-        Set some known filters on the jinja environment
-        """
         filters = self.jinja_env.filters
-        filters['markdown'] = template_filters.markdown
-        filters['media_url'] = template_filters.media_url_func(self.site)
-
-    def ensure_output_path(self):
-        if not os.path.exists(self.output_path):
-            os.makedirs(self.output_path)
+        filters['markdown'] = markdown
+        filters['media_url'] = media_url_func(self.site)
 
     def render_file(self, path, template, **kwargs):
         with open(path, 'w') as f:
@@ -68,36 +59,6 @@ class Renderer(object):
 
         self.render_site_posts()
         self.render_site_tags()
-        self.render_site_json()
-
-    def render_site_json(self):
-        self.ensure_output_path()
-
-        json_path = os.path.join(self.output_path, 'site.json')
-        payload = {
-            'pages': [],
-            'posts': [],
-        }
-
-        for page in self.site.payload['pages']:
-            p = page.as_dict()
-            p['body'] = template_filters.markdown(p['body'])
-            payload['pages'].append(p)
-
-        for post in self.site.payload['posts']:
-            p = post.as_dict()
-            p['blurb'] = template_filters.markdown(p['blurb'])
-            p['body'] = template_filters.markdown(p['body'])
-            p['date'] = post['date'].isoformat()
-            payload['posts'].append(p)
-
-        for k, v in self.site.payload.items():
-            if k not in {'posts', 'pages'}:
-                payload[k] = v
-
-        # markdown-render each post and page
-        with open(json_path, 'w') as f:
-            f.write(json.dumps(payload))
 
     def render_posts(self, posts, prefix=''):
         """
