+++ posty/renderer/rss.py
@@ -1,0 +1,5 @@
+from .base import Renderer
+
+
+class RssRenderer(Renderer):
+    pass
