+++ tests/renderer/test_json.py
@@ -1,0 +1,28 @@
+import json
+import os
+import pytest
+
+from posty.renderer import JsonRenderer
+
+from ..fixtures import site   # noqa
+
+
+@pytest.fixture     # noqa
+def renderer(site):
+    site.load()
+    return JsonRenderer(site)
+
+
+def test_render_site(renderer):     # noqa
+    """
+    Verify that Site.render() spits out a valid JSON file
+    """
+    renderer.render_site()
+
+    json_path = os.path.join(renderer.output_path, 'site.json')
+    blob = json.load(open(json_path))
+    assert blob['title'] == 'Test website'
+
+    assert len(blob['pages']) > 0
+    assert len(blob['posts']) > 0
+    assert len(blob['tags']) > 0
