+++ tests/renderer/test_util.py
@@ -1,6 +1,6 @@
-from posty import template_filters
+from posty.renderer import util
 
-from .fixtures import site  # noqa
+from ..fixtures import site  # noqa
 
 
 def test_markdown():
@@ -8,10 +8,10 @@ def test_markdown():
     Really basic test. No need to test markdown itself, just our use of it
     """
     fenced = "```\nfarts.\n```"
-    result = template_filters.markdown(fenced)
+    result = util.markdown(fenced)
     assert result == "<pre><code>farts.\n</code></pre>"
 
 
 def test_media_url_func(site):  # noqa
-    func = template_filters.media_url_func(site)
+    func = util.media_url_func(site)
     assert func('jawn') == '/test/media/jawn'
